POC/requirements.txt
selenium
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
openpyxl
pytest-ordering
parameterized
requests
beautifulsoup4
allure-pytest